homework 5

+++ src/services/contacts.js
@@ -9,11 +9,12 @@ export const getAllContacts = async ({
   sortOrder = SORT_ORDER.ASC,
   sortBy = '_id',
   filter = {},
+  userId,
 }) => {
   const limit = perPage;
   const skip = (page - 1) * perPage;
 
-  const contactsQuery = contactsCollection.find();
+  const contactsQuery = contactsCollection.find({ userId });
 
   if (filter.contactType) {
     contactsQuery.where('contactType').equals(filter.contactType);
@@ -41,8 +42,8 @@ export const getAllContacts = async ({
   };
 };
 
-export const getContactsById = async (contactId) => {
-  const contact = await contactsCollection.findById(contactId);
+export const getContactsById = async ({ _id, userId }) => {
+  const contact = await contactsCollection.findOne({ _id, userId });
   return contact;
 };
 
@@ -51,16 +52,19 @@ export const createContact = async (payload) => {
   return contact;
 };
 
-export const deleteContact = async (contactId) => {
-  const contact = await contactsCollection.findOneAndDelete({
-    _id: contactId,
-  });
+export const deleteContact = async ({ _id, userId }) => {
+  const contact = await contactsCollection.findOneAndDelete({ _id, userId });
   return contact;
 };
 
-export const updateContact = async (contactId, payload, options = {}) => {
+export const updateContact = async (
+  contactId,
+  payload,
+  userId,
+  options = {},
+) => {
   const rawResult = await contactsCollection.findOneAndUpdate(
-    { _id: new ObjectId(contactId) },
+    { _id: new ObjectId(contactId), userId },
     { $set: payload },
     {
       returnDocument: 'after',
